+++ Modules/Intent.Modules.AspNetCore.IntegrationTesting/NugetPackages.cs
@@ -225,7 +225,7 @@ namespace Intent.Modules.AspNetCore.IntegrationTesting
             NugetRegistry.Register(XunitRunnerVisualstudioPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("3.1.4"),
+                        ( >= 8, >= 0) => new PackageVersion("3.1.5"),
                         ( >= 6, >= 0) => new PackageVersion("3.0.2"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{XunitRunnerVisualstudioPackageName}'"),
                     }
@@ -233,10 +233,10 @@ namespace Intent.Modules.AspNetCore.IntegrationTesting
             NugetRegistry.Register(XunitV3PackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("3.0.1")
+                        ( >= 8, >= 0) => new PackageVersion("3.1.0")
                             .WithNugetDependency("xunit.analyzers", "1.24.0")
-                            .WithNugetDependency("xunit.v3.assert", "3.0.1")
-                            .WithNugetDependency("xunit.v3.core", "3.0.1"),
+                            .WithNugetDependency("xunit.v3.assert", "3.1.0")
+                            .WithNugetDependency("xunit.v3.core", "3.1.0"),
                         ( >= 6, >= 0) => new PackageVersion("1.1.0")
                             .WithNugetDependency("xunit.analyzers", "1.20.0")
                             .WithNugetDependency("xunit.v3.assert", "1.1.0")