+++ Modules/Intent.Modules.UnitTesting/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.1
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 1.0.0
 
 - Improvement: Updated NuGet package versions.